fix pathing

+++ deependeliminator/standings.py
@@ -71,4 +71,4 @@ def get_week():
     opening_day = datetime.datetime.strptime(os.environ.get('OPENING_DAY', '2018-09-06'), '%Y-%m-%d')
     now = datetime.datetime.now()
 
-    return ((now - opening_day).days / 7) + 1
+    return int(((now - opening_day).days / 7) + 1)
